requirements.txt
genai-apis[gemini]
genai-apis[opeani]
genai-apis[anthropic]
networkx[default]
pydot
pyyaml
futures
GitPython